Rename configurators to spec handlers
